take FileCrypto class out from uploader.py
